+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+__2017-04-06__
+
+  + app.yaml 不再支持 `binds`, `mount_path`, `permdir` 这几个关键字, 统一用 `volume` 来代替, 详见[app.yaml 说明](docs/user-docs/specs.md)
+
 __2017-04-01__
 
   + 文档开始迁移到 gitlab pages